smart_alarm/requirements/requirements.txt
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.7
django-environ==0.4.4
django-rest-framework-social-oauth2==1.0.8
djangorestframework==3.7.3
idna==2.6
psycopg2==2.7.3.2
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.0.0
urllib3==1.22
